+++ packages/terminal/src/node/shell-terminal-server.ts
@@ -16,15 +16,18 @@
 
 import { inject, injectable, named } from '@theia/core/shared/inversify';
 import { ILogger } from '@theia/core/lib/common/logger';
+import { EnvironmentUtils } from '@theia/core/lib/node/environment-utils';
 import { IShellTerminalServerOptions } from '../common/shell-terminal-protocol';
 import { BaseTerminalServer } from './base-terminal-server';
-import { mergeProcessEnv, ShellProcessFactory } from './shell-process';
+import { ShellProcessFactory } from './shell-process';
 import { ProcessManager } from '@theia/process/lib/node';
 import { isWindows } from '@theia/core/lib/common/os';
 import * as cp from 'child_process';
 
 @injectable()
 export class ShellTerminalServer extends BaseTerminalServer {
+
+    @inject(EnvironmentUtils) protected environmentUtils: EnvironmentUtils;
 
     constructor(
         @inject(ShellProcessFactory) protected readonly shellFactory: ShellProcessFactory,
@@ -35,7 +38,7 @@ export class ShellTerminalServer extends BaseTerminalServer {
 
     async create(options: IShellTerminalServerOptions): Promise<number> {
         try {
-            options.env = this.mergeProcessEnv(options.env);
+            options.env = this.environmentUtils.mergeProcessEnv(options.env);
             this.mergedCollection.applyToProcessEnvironment(options.env);
             const term = this.shellFactory(options);
             this.postCreate(term);
@@ -44,15 +47,6 @@ export class ShellTerminalServer extends BaseTerminalServer {
             this.logger.error('Error while creating terminal', error);
             return -1;
         }
-    }
-
-    /**
-     * Empty string values will be removed from the final env.
-     *
-     * @param env desired environment to merge with `process.env`.
-     */
-    protected mergeProcessEnv(env: Record<string, string | null> = {}): Record<string, string> {
-        return mergeProcessEnv(env);
     }
 
     // copied and modified from https://github.com/microsoft/vscode/blob/4636be2b71c87bfb0bfe3c94278b447a5efcc1f1/src/vs/workbench/contrib/debug/node/terminals.ts#L32-L75